requirements.txt
streamlit>=1.24.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
lxml>=4.9.0
